+++ packages/example-v8/src/stories/WithTypedProps.spec.ts
@@ -1,6 +1,6 @@
 import {test} from "@playwright/test";
 import {assertStorybookIsRunning} from "../tests/utils";
-import StorybookPageObject from "../tests/utils/StorybookPage";
+import StorybookPageObject from "../tests/utils/AppObject";
 import {TEST_TIMEOUT_MS} from "../tests/utils/constants";
 
 test.beforeAll(assertStorybookIsRunning);
@@ -12,7 +12,10 @@ test.beforeEach(async ({page}) => {
 
 test("shows default controls when initial values are not defined", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
-  await storybookPage.action.openStoriesTreeItemById("story", "stories-withtypedprops--default-enabled");
+  await storybookPage.action.openStoriesTreeItemById(
+    "story",
+    "stories-withtypedprops--default-enabled",
+  );
 
   const storyPage = storybookPage.activeStoryPage;
   await storyPage.argsTable.assert.controlsMatch({
@@ -61,7 +64,10 @@ test("shows deep controls when initial values are defined", async ({page}) => {
 
 test("supports customising controls with initial values", async ({page}) => {
   const storybookPage = new StorybookPageObject(page);
-  await storybookPage.action.openStoriesTreeItemById("story", "stories-withtypedprops--with-custom-controls");
+  await storybookPage.action.openStoriesTreeItemById(
+    "story",
+    "stories-withtypedprops--with-custom-controls",
+  );
 
   const storyPage = storybookPage.activeStoryPage;
   await storyPage.argsTable.assert.controlsMatch({
